Work around problem with empty database EPG entries
Kodi does not create EPG entries for channels that do not exist when the
PVR manager is started. As a result, when the database is new/empty the
user will not get any EPG data until Kodi is restarted.  This workaround
executes synchronous discovery tasks in ADDON_Create to seed the
database with information that can be read by Kodi immediately on
startup