Fix problem with date/time only timers temporarily disappearing immediately after scheduling
When scheduling a date/time only rule the indicator would appear and
then disappear.  Refreshing the guide and recording rule data would
bring it back.  The problem was timing related -- the update to the EPG
was being triggered before the update to the timers/timer rules so Kodi
was clearing the indicator since there was no timer/timer rule in place
at the time of the EPG update