Switch to Visual C++ 2013
Jarvis does not ship with the Visual C++ 2015 runtime library; move the
Jarvis code back to Visual Studio 2013 and apply the required changes
for compatibility with that Platform Toolset version.  Note that there
is a known defect in VS2013 when _USE_32BIT_TIME_T is defined for the
project in conjunction with using the std::chrono namespace. Code that
requires the use of a std::chrono namespace wait operation had to be
moved into separate translation units with no pre-compiled header
support.