Fix defects in enumerate_timers
Two defects were fixed: first, when a DateTimeOnly and Series rule
overlapped on a single episode a Series Episode timer was always
created.  The parent recordingruleid was not being retrieved correctly.
Second, the PVR Client was intended to expose ALL episodes of a Series
rule but was only exposing one.  This turned out to be an old copy/paste
error from a while ago that was sending Kodi the channelid rather than
generating a unique timerid.  Therefore only one episode would show up
for any given channel.  The original behavior of generating a unique
timerid has been restored.