Solve VS2013 condition_variable problem a different way
Ran into this a third time after the recent changes and opted to create
a helper function in it's own translation unit rather than keeping the
affected code isolated from stdafx.h.  Trying to keep the utility code
(livestream, database, scheduler, etc) for Jarvis and Krypton as close
to the same as possible.