Fix phantom channel problem when legacy device is detected
Legacy devices with newer firmware are being detected from
my.hdhomerun.com and passing through enough code to be listed in the
database.  These devices use an online URL to acquire lineup data, which
(for me) always returns the string 'null'.  When the lineup is
generated, this phantom entry is being passed into Kodi as a channel 0
with no additional information, which is skewing the channel numbers
among other things.  Exclude the legacy devices from the search, but
also watch out for that 'null' string when generating the lineup.